# packages/shared-types/api.ts
// Generated by scripts/export_openapi.py. Do not edit; `task.py openapi-check` gates it.

export interface AnswerClaim {
  text: string;
  metric_id: string;
  value: number | string;
  unit: string;
  evidence_id: string;
}

export interface DataProvenance {
  transaction_records: string;
  aggregate_calibration: string;
  sources_document: string;
  scenario_id: string;
  seed: number;
  ground_truth: string;
  parameter_counts: ParameterCounts;
  fee_schedule: FeeRuleView[];
  checksums: Record<string, string>;
  disclaimer: string;
}

export interface EvidenceDetail {
  execution_id: string;
  evidence_id: string;
  metric_id: string;
  unit: string;
  value: number | string;
  period_from: string;
  period_to: string;
  dimension_value: string | null;
  inputs: Record<string, number | string>;
  rules_applied: string[];
  verification_checks: string[];
  provenance: ProvenanceLevel;
  source_record_ids: string[];
}

export interface EvidenceIndex {
  execution_id: string;
  items: EvidenceLine[];
}

export interface EvidenceLine {
  evidence_id: string;
  tool_name: string;
  tool_version: string;
  metric_id: string;
  unit: string;
  value: number | string;
  period_from: string;
  period_to: string;
  dimension_value: string | null;
  support: "FORMULA" | "AGGREGATION";
}

export interface ExceptionItem {
  id: string;
  category: string;
  side: string;
  transaction_id: string | null;
  settlement_id: string | null;
  amount_paise: number;
  currency?: string;
  status: string;
  detail: Record<string, unknown>;
}

export interface ExceptionPage {
  items: ExceptionItem[];
  next_cursor?: string | null;
}

export interface ExecutionLine {
  execution_id: string;
  merchant_id: string;
  question: string;
  status: string;
  response_source: string | null;
  created_at: string;
  period_from: string | null;
  period_to: string | null;
}

export interface ExecutionPage {
  items: ExecutionLine[];
  next_cursor: string | null;
}

export interface ExecutionSummary {
  execution_id: string;
  merchant_id: string;
  period_from: string | null;
  period_to: string | null;
  status: string;
  response_source: string | null;
  answer: string | null;
  claims: AnswerClaim[];
  grounding_attempts: number;
  error: Record<string, unknown> | null;
}

export interface FeeRuleView {
  instrument: string;
  mdr_rate: string;
  platform_fee_rate: string;
  threshold_paise: number;
  flat_fee_paise: number;
  provenance: string;
  note: string;
}

export interface HTTPValidationError {
  detail?: ValidationError[];
}

export interface Health {
  status: "ok";
  version: string;
  environment: string;
}

export interface MatchRecord {
  id: string;
  run_id: string;
  rule: string;
  confidence_ratio: string;
  reason: string;
  amount_delta_paise: number;
  lag_days: number;
  transaction: Record<string, unknown>;
  settlement: Record<string, unknown>;
}

export interface ParameterCounts {
  CITED?: number;
  DERIVED?: number;
  ASSUMED?: number;
}

export interface Period {
  from: string;
  to: string;
}

export interface ProvenanceLevel {
  evidence_id: string;
  tool_name: string;
  tool_version: string;
  metric_id: string;
  unit: string;
  value: number | string;
  period_from: string;
  period_to: string;
  dimension_value: string | null;
  support: "FORMULA" | "AGGREGATION";
  detail: string;
  rules_applied: string[];
  operands: ProvenanceOperand[];
  source_record_ids: string[];
}

export interface ProvenanceOperand {
  name: string;
  reference: string;
  value: number | string;
  node: ProvenanceLevel | null;
}

export interface RunAccepted {
  execution_id: string;
  status: string;
  replayed?: boolean;
}

export interface RunPage {
  items: RunSummary[];
  next_cursor?: string | null;
}

export interface RunRequest {
  merchant_id: string;
  message: string;
  client_request_id?: string | null;
}

export interface RunSummary {
  run_id: string;
  period: Period;
  ledger_count: number;
  bank_count: number;
  matched_pairs_count: number;
  matched_clean_count: number;
  matched_with_exception_count: number;
  unmatched_ledger_count: number;
  unmatched_bank_count: number;
  clean_match_rate_ratio: string;
  exception_count: number;
  exception_breakdown: Record<string, number>;
  unresolved_exception_value_paise: number;
}

export interface ValidationError {
  loc: (string | number)[];
  msg: string;
  type: string;
}

export type ApiPath =
  | "/api/v1/agent/runs"
  | "/api/v1/agent/runs/{execution_id}/events"
  | "/api/v1/executions"
  | "/api/v1/executions/{execution_id}"
  | "/api/v1/executions/{execution_id}/evidence"
  | "/api/v1/executions/{execution_id}/evidence/{evidence_id}"
  | "/api/v1/health"
  | "/api/v1/provenance"
  | "/api/v1/reconciliation/runs"
  | "/api/v1/reconciliation/runs/{run_id}/exceptions"
  | "/api/v1/reconciliation/runs/{run_id}/matches/{match_id}"
  | "/health";