# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunAccepted": {
        "properties": {
          "execution_id": {
            "title": "Execution Id",
            "type": "string"
          },
          "replayed": {
            "default": false,
            "title": "Replayed",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "execution_id",
          "status"
        ],
        "title": "RunAccepted",
        "type": "object"
      },
...
      "RunRequest": {
        "description": "What a client asks for. ``merchant_id`` is checked, never trusted.",
        "properties": {
          "client_request_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Request Id"
          },
          "merchant_id": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Merchant Id",
            "type": "string"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "merchant_id",
          "message"
        ],
        "title": "RunRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Agentic financial computation and reconciliation. The LLM decides what to compute; deterministic code decides what the number is.",
    "title": "RazorMind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agent/runs": {
      "post": {
        "description": "Reserve an execution, start it in the background, and return its id.\n\nThe row is inserted *before* returning, not by the background task: a\nclient that polls the id it was just handed must find something there, and\n\"202 Accepted, but ask again in a moment\" is an API that leaks a race.",
        "operationId": "start_run_api_v1_agent_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-razormind-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Razormind-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Run",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/agent/runs/{execution_id}/events": {
      "get": {
        "description": "Replay from the log, then follow the live broadcaster.",
        "operationId": "stream_events_api_v1_agent_runs__execution_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from_seq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "From Seq",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Events",
        "tags": [
          "agent"
        ]
      }
    },
...
  }
}